adhd-dashboard/requirements.txt
# ============================================
# 🧠 ADHD Dashboard v3.0 - Python Dependencies
# ============================================
# نصب: pip install -r requirements.txt --break-system-packages
# ============================================

# Flask Framework
Flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
Werkzeug>=3.0.0

# Notion API
notion-client>=2.2.0

# Google Sheets API
gspread>=5.12.0
gspread-formatting>=1.1.2
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.111.0

# Utilities
requests>=2.31.0

# Optional: Telegram Bot (فاز بعدی)
# python-telegram-bot>=20.7
